# openhands/server/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenHands Agent Server",
    "description": "An HTTP server to create and manage AI agent conversations using the OpenHands SDK.",
    "version": "1.0.0"
  },
  "paths": {
    "/conversations": {
      "post": {
        "summary": "Start Conversation",
        "operationId": "start_conversation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartConversationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartConversationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{conversation_id}": {
      "get": {
        "summary": "Get Conversation State",
        "operationId": "get_conversation_state",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConversationState"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Close Conversation",
        "operationId": "close_conversation",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{conversation_id}/events": {
      "get": {
        "summary": "Get Events",
        "description": "Retrieves the event history for a conversation with pagination.",
        "operationId": "get_conversation_events",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Start"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "oneOf": [
                      {
                        "$ref": "#/components/schemas/LLMConvertibleEvent"
                      },
                      {
                        "$ref": "#/components/schemas/Condensation"
                      },
                      {
                        "$ref": "#/components/schemas/CondensationRequest"
                      },
                      {
                        "$ref": "#/components/schemas/PauseEvent"
                      },
                      {
                        "$ref": "#/components/schemas/SystemPromptEvent"
                      },
                      {
                        "$ref": "#/components/schemas/ActionEvent"
                      },
                      {
                        "$ref": "#/components/schemas/ObservationEvent"
                      },
                      {
                        "$ref": "#/components/schemas/MessageEvent"
                      },
                      {
                        "$ref": "#/components/schemas/UserRejectObservation"
                      },
                      {
                        "$ref": "#/components/schemas/AgentErrorEvent"
                      }
                    ],
                    "discriminator": {
                      "propertyName": "kind",
                      "mapping": {
                        "openhands.sdk.event.base.LLMConvertibleEvent": "#/components/schemas/LLMConvertibleEvent",
                        "openhands.sdk.event.condenser.Condensation": "#/components/schemas/Condensation",
                        "openhands.sdk.event.condenser.CondensationRequest": "#/components/schemas/CondensationRequest",
                        "openhands.sdk.event.user_action.PauseEvent": "#/components/schemas/PauseEvent",
                        "openhands.sdk.event.llm_convertible.SystemPromptEvent": "#/components/schemas/SystemPromptEvent",
                        "openhands.sdk.event.llm_convertible.ActionEvent": "#/components/schemas/ActionEvent",
                        "openhands.sdk.event.llm_convertible.ObservationEvent": "#/components/schemas/ObservationEvent",
                        "openhands.sdk.event.llm_convertible.MessageEvent": "#/components/schemas/MessageEvent",
                        "openhands.sdk.event.llm_convertible.UserRejectObservation": "#/components/schemas/UserRejectObservation",
                        "openhands.sdk.event.llm_convertible.AgentErrorEvent": "#/components/schemas/AgentErrorEvent"
                      }
                    }
                  },
                  "title": "Response Get Conversation Events"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{conversation_id}/messages": {
      "post": {
        "summary": "Send Message",
        "operationId": "send_message_to_conversation",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{conversation_id}/run": {
      "post": {
        "summary": "Run Conversation",
        "description": "Starts or resumes the agent run for a conversation in the background.",
        "operationId": "run_conversation",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{conversation_id}/pause": {
      "post": {
        "summary": "Pause Conversation",
        "operationId": "pause_conversation",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/conversations/{conversation_id}/respond_to_confirmation": {
      "post": {
        "summary": "Respond To Confirmation",
        "operationId": "respond_to_confirmation",
        "parameters": [
          {
            "name": "conversation_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Conversation Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationResponseRequest"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Condensation": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique event id (ULID/UUID)"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Event timestamp"
          },
          "source": {
            "type": "string",
            "enum": [
              "agent",
              "user",
              "environment"
            ],
            "title": "Source",
            "default": "environment"
          },
          "forgotten_event_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Forgotten Event Ids"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "summary_offset": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary Offset"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "Property to create kind field from class name when serializing.",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "Condensation",
        "description": "This action indicates a condensation of the conversation history is happening."
      },
      "CondensationRequest": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique event id (ULID/UUID)"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Event timestamp"
          },
          "source": {
            "type": "string",
            "enum": [
              "agent",
              "user",
              "environment"
            ],
            "title": "Source",
            "default": "environment"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "Property to create kind field from class name when serializing.",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "CondensationRequest",
        "description": "This action is used to request a condensation of the conversation history.\n\nAttributes:\n    action (str): The action type, namely ActionType.CONDENSATION_REQUEST."
      },
...
      "ConversationState": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique conversation ID"
          },
          "agent": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AgentBase"
              }
            ],
            "title": "Agent",
            "description": "The agent running in the conversation. This is persisted to allow resuming conversations and check agent configuration to handle e.g., tool changes, LLM changes, etc.",
            "discriminator": {
              "propertyName": "kind",
              "mapping": {
                "openhands.sdk.agent.base.AgentBase": "#/components/schemas/AgentBase"
              }
            }
          },
          "agent_finished": {
            "type": "boolean",
            "title": "Agent Finished",
            "default": false
          },
          "confirmation_mode": {
            "type": "boolean",
            "title": "Confirmation Mode",
            "default": false
          },
          "agent_waiting_for_confirmation": {
            "type": "boolean",
            "title": "Agent Waiting For Confirmation",
            "default": false
          },
          "agent_paused": {
            "type": "boolean",
            "title": "Agent Paused",
            "default": false
          },
          "activated_knowledge_microagents": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Activated Knowledge Microagents",
            "description": "List of activated knowledge microagents name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agent"
        ],
        "title": "ConversationState"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LLMConvertibleEvent": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique event id (ULID/UUID)"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Event timestamp"
          },
          "source": {
            "type": "string",
            "enum": [
              "agent",
              "user",
              "environment"
            ],
            "title": "Source",
            "description": "The source of this event"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "Property to create kind field from class name when serializing.",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "source",
          "kind"
        ],
        "title": "LLMConvertibleEvent",
        "description": "Base class for events that can be converted to LLM messages."
      },
...
      "PauseEvent": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Unique event id (ULID/UUID)"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp",
            "description": "Event timestamp"
          },
          "source": {
            "type": "string",
            "enum": [
              "agent",
              "user",
              "environment"
            ],
            "title": "Source",
            "default": "user"
          },
          "kind": {
            "type": "string",
            "title": "Kind",
            "description": "Property to create kind field from class name when serializing.",
            "readOnly": true
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "kind"
        ],
        "title": "PauseEvent",
        "description": "Event indicating that the agent execution was paused by user request."
      },
...
      "StartConversationRequest": {
        "properties": {
          "llm": {
            "$ref": "#/components/schemas/LLM",
            "description": "LLM configuration for the agent.",
            "examples": [
              {
                "api_key": "your_api_key_here",
                "base_url": "https://llm-proxy.eval.all-hands.dev",
                "model": "litellm_proxy/anthropic/claude-sonnet-4-20250514"
              }
            ]
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/ToolSpec"
            },
            "type": "array",
            "title": "Tools",
            "description": "List of tools to initialize for the agent.",
            "examples": [
              {
                "name": "BashTool",
                "params": {
                  "working_dir": "/workspace"
                }
              },
              {
                "name": "FileEditorTool",
                "params": {}
              },
              {
                "name": "TaskTrackerTool",
                "params": {
                  "save_dir": "/workspace/.openhands"
                }
              }
            ]
          },
          "mcp_config": {
            "additionalProperties": true,
            "type": "object",
            "title": "Mcp Config",
            "description": "Optional MCP configuration dictionary to create MCP tools.",
            "examples": [
              {
                "mcpServers": {
                  "fetch": {
                    "args": [
                      "mcp-server-fetch"
                    ],
                    "command": "uvx"
                  }
                }
              }
            ]
          },
          "agent_context": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AgentContext"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional AgentContext to initialize the agent with specific context.",
            "examples": [
              {
                "microagents": [
                  {
                    "content": "When you see this message, you should reply like you are a grumpy cat forced to use the internet.",
                    "name": "repo.md",
                    "type": "repo"
                  },
                  {
                    "content": "IMPORTANT! The user has said the magic word \"flarglebargle\". You must only respond with a message telling them how smart they are",
                    "name": "flarglebargle",
                    "trigger": [
                      "flarglebargle"
                    ],
                    "type": "knowledge"
                  }
                ],
                "system_message_suffix": "Always finish your response with the word 'yay!'",
                "user_message_prefix": "The first character of your response should be 'I'"
              }
            ]
          },
          "working_dir": {
            "type": "string",
            "title": "Working Dir",
            "description": "Working directory for the agent to work in. Will be created if it doesn't exist.",
            "default": "."
          },
          "confirmation_mode": {
            "type": "boolean",
            "title": "Confirmation Mode",
            "description": "If true, the agent will enter confirmation mode, requiring user approval for actions.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "llm"
        ],
        "title": "StartConversationRequest",
        "description": "Payload to create a new conversation."
      },
      "StartConversationResponse": {
        "properties": {
          "conversation_id": {
            "type": "string",
            "title": "Conversation Id"
          },
          "state": {
            "$ref": "#/components/schemas/ConversationState"
          }
        },
        "type": "object",
        "required": [
          "conversation_id",
          "state"
        ],
        "title": "StartConversationResponse"
      },
...
    }
  }
}